String Literal Addition to Regex
Looks like the https regex needed the raw string literal character on it (r'). No clue if this will show up with tab or spaces, just using the online editor on it. Forgive me, PEP-8.

Tested on two comments that broke the bot with the online compiler/interpreter here: https://repl.it/JJTo/1